+++ NEWS.md
@@ -1,2 +1,3 @@
 <!-- Changes in v1.0a1 that will not backport to v0.9 and are not related to tests or documentation -->
 - The default value of the `sortby` parameter in `pep265.itemlist` will become `VALUES` instead of `ORIGINAL` in future versions; not specifying it now issues a `DeprecationWarning`.
+- `pep294.apply()` now returns the modified module.